Get rid of the client-side sha1 "pre-hash".
Even if sha1 would still be considered cryptographically secure: I do
not thing that it is a good idea to allow logging in by just providing
the hash of the password. This does not secure the login process: an
attacker hijacking the password hash can just login with that hash. So
we can as well submit the password in clear text.

Also: pre hashing makes it impossible to use external authentication
methods like LDAP or anything else.